requirements.txt
# Build dependencies
click==6.7
future>=0.15.2
gitpython==2.1.1
pandas==0.19.2
pyyaml==3.12
radon==1.4.2
scandir==1.4
six==1.10.0
tqdm==4.11.2

# Test dependencies
flake8==2.6.2
pytest==3.0.6
tox==2.6.0
